Preps for rewriting plot
