+++ pyioflash/__init__.py
@@ -8,7 +8,7 @@ from pyioflash.simulation import SimulationData
 from pyioflash.simulation import NameData
 
 # Used to create professional 2D plots
-from pyioflash.visual import SimulationPlot
+#from pyioflash.visual import SimulationPlot
 from pyioflash.visual import FigureOptions, PlotOptions, AnimationOptions
 from pyioflash.visual import Plane
