+++ pyioflash/visual/__init__.py
@@ -4,10 +4,10 @@ and process FLASH4 HDF5 plt and chk files
 """
 
 # Used to import and handle simulation data
-from pyioflash.visual.plot import SimulationPlot
+#from pyioflash.visual.plot import SimulationPlot
 
 # Used to assist in defining visualization options
 from pyioflash.visual.options import FigureOptions, PlotOptions, AnimationOptions
 
 # Used to assist in defining ploting constructs
-from pyioflash.visual.geometry import Plane
+from pyioflash.visual.primitive import Plane
