+++ pyioflash/visual/primitive.py
